refactor(scheduler): split scheduler.go into stage_runner.go and budget.go
Split the 791-line scheduler.go into three files under the 500-line limit:

- scheduler.go (319 lines): core types, Run(), pollAndExecute(), executeRun()

- stage_runner.go (207 lines): executeStage(), executeStep(), markStepFailed()

- budget.go (293 lines): ErrBudgetExceeded, resolveStepLimits(),

  recordStepUsage(), resolveStepTokenData(), buildStepMetadata(),

  resolveAgentModel(), parseMetadataInt64()

Pure refactor — no behavioral changes. All 2211 test lines pass with -race.

Package external surface unchanged (only cmd/orchestrator/main.go imports it).

v0.2.0 Release Preparation Plan PR 2 / Track A.
