+++ FILEMAP.md
@@ -1,13 +1,13 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-17  
-> **227** tracked files
+> **229** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
-| `internal/` | 50 |
+| `internal/` | 52 |
 | `agents/` | 39 |
 | `docs/` | 34 |
 | `tests/` | 30 |
@@ -28,7 +28,7 @@
 |-----------|-------|
 | `.py` | 78 |
 | `.md` | 49 |
-| `.go` | 46 |
+| `.go` | 48 |
 | `.yaml` | 22 |
 | `.rs` | 7 |
 | `(no extension)` | 5 |
@@ -252,8 +252,10 @@
 │   ├── resilience/
 │   │   └── resilience.go
 │   ├── scheduler/
+│   │   ├── budget.go
 │   │   ├── scheduler.go
-│   │   └── scheduler_test.go
+│   │   ├── scheduler_test.go
+│   │   └── stage_runner.go
 │   ├── security/
 │   │   └── security.go
 │   ├── server/
