+++ docs/v0.2-release-prep-plan.md
@@ -18,8 +18,8 @@ Update this table as each PR merges. Flip **Status** and add GitHub PR number + 
 
 | PR | Track | Title | Branch | Status | GitHub PR | Merged |
 |----|-------|-------|--------|--------|-----------|--------|
-| 1 | A | Split `agents/persona_runtime.py` | `feature/v02-release-prep-split-persona-runtime` | 🔀 PR open | #95 | — |
-| 2 | A | Split `internal/scheduler/scheduler.go` | `feature/v02-release-prep-split-scheduler` | ⬜ Not started | — | — |
+| 1 | A | Split `agents/persona_runtime.py` | `feature/v02-release-prep-split-persona-runtime` | ✅ Merged | #95 | 2026-04-17 |
+| 2 | A | Split `internal/scheduler/scheduler.go` | `feature/v02-release-prep-split-scheduler` | 🔀 PR open | — | — |
 | 3 | A | Split `agents/memory/episodic.py` + `agents/server.py` | `feature/v02-release-prep-split-memory-server` | ⬜ Not started | — | — |
 | 4 | B | README overhaul | `feature/v02-release-prep-readme` | ⬜ Not started | — | — |
 | 5 | B | Persona & memory user guide | `feature/v02-release-prep-persona-guide` | ⬜ Not started | — | — |
@@ -104,7 +104,7 @@ Splitting oversized files before docs/tests ensures docs can reference stable mo
 
 ### PR 1 — Split `agents/persona_runtime.py`
 
-**Status**: 🔀 PR open — PR #95  
+**Status**: ✅ Merged — PR #95  
 **Branch**: `feature/v02-release-prep-split-persona-runtime`  
 **Depends on**: nothing  
 **Estimated delta**: ~0 net lines (refactor only)
@@ -133,7 +133,7 @@ Splitting oversized files before docs/tests ensures docs can reference stable mo
 
 ### PR 2 — Split `internal/scheduler/scheduler.go`
 
-**Status**: ⬜ Not started  
+**Status**: 🔀 PR open  
 **Branch**: `feature/v02-release-prep-split-scheduler`  
 **Depends on**: nothing  
 **Estimated delta**: ~0 net lines (refactor only)
